+++ internal/tests/GroupDataSource/group_data_source_test.go
@@ -21,7 +21,7 @@ func TestGroupDataSource(t *testing.T) {
 					resource.TestCheckResourceAttr("data.axual_group.team-integrations-imported", "email_address", "test.user@axual.com"),
 					resource.TestCheckResourceAttr("data.axual_group.team-integrations-imported", "phone_number", "+6112356789"),
 					resource.TestCheckResourceAttr("data.axual_group.team-integrations-imported", "members.#", "1"),
-					resource.TestCheckResourceAttrPair("data.axual_group.team-integrations-imported", "members.0", "axual_user.bob", "id"),
+					resource.TestCheckResourceAttrPair("data.axual_group.team-integrations-imported", "members.0", "data.axual_user.test_user", "id"),
 				),
 			},
 			{