+++ internal/tests/GroupResource/axual_group_initial.tf
@@ -1,23 +1,11 @@
-resource "axual_user" "bob" {
-  first_name    = "Bob"
-  last_name     = "Foo"
-  email_address = "bob.foo@example.com"
-  phone_number  = "+123456"
-  roles = [
-    { name = "APPLICATION_AUTHOR" },
-    { name = "ENVIRONMENT_AUTHOR" },
-    { name = "STREAM_AUTHOR" }
-  ]
-}
-
 resource "axual_group" "team-integrations" {
   name          = "testgroup9999"
   phone_number  = "+6112356789"
   email_address = "test.user@axual.com"
   members = [
-    axual_user.bob.id,
+    data.axual_user.test_user.id,
   ]
   managers = [
-    axual_user.bob.id,
+    data.axual_user.test_user.id,
   ]
 }