+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 All notable changes to this project will be documented in this file.
 
 ## [master](https://github.com/Axual/terraform-provider-axual/blob/master) - TBR
+* Add User data source
 
-## [2.5.5](https://github.com/Axual/terraform-provider-axual/releases/tag/v2.6.4) - 2025-03-07
+## [2.5.5](https://github.com/Axual/terraform-provider-axual/releases/tag/v2.5.5) - 2025-03-07
 * Removed unused method for `/groups/{uid}/members/{uid}`
 * Auth0 authentication support for Axual Trial Environment